pyproject.toml
[project]
name = "caption-mate"
version = "0.1.0"
description = "A search and download tool for captions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0.0",
    "smbprotocol>=1.12.0",
    "pysmb>=1.2.9",
    "ffmpeg-python>=0.2.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
caption-mate = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
